==> scripts/Daniel/scripts/01_PrepareEdges.ipynb <==
# Preparing edge table

## Summary

- Import node table from Natalie.
- Extract list of traits.
- Identify sources of GWAS for the traits.
- Save table with links to test, specifying whether a cis-analysis or a genome-wide analysis is needed.

## Libraries

library(readr)
library(dplyr)
library(tidyr)
library(purrr)
library(ggplot2)

projfold <- "/Users/da1078co/Documents/Lund/PhD/Projects/BN"

## Edges to test - from Natalie

edge_tb0 <- read_tsv(
  file.path(projfold, "data", "edge_table_original.tsv"),
  show_col_types = FALSE
)

head(edge_tb0)

## List of traits

trait_tb0 <- tibble(
  trait = c(
    edge_tb0[, 1, drop = TRUE],
    edge_tb0[, 2, drop = TRUE]
  )
) |>
  unique() |>
  arrange(trait)

trait_tb0$trait

## Gene traits

gene_traits <- c(
  "AGRP", "APOM", "CD4", "CDH5", "CTRC", "CTSD", "FAM3C",
  "FGF21", "HMOX1", "IGFBP1", "IGFBP2", "KITLG", "LDLR",
  "LEP", "LPL", "MATN2", "MFGE8", "NRP1", "PON3", "TGM2"
)

### GWAS source

We will use UKB-PPP summary statistics for these genes.

The process to download the files corresponding to these genes is described in:

- `03a_Download_pQTLPanel.py`
- `03b_Process_pQTLPanel.ipynb`
- `03c_Download_pQTL.ipynb`

### Gencode table to obtain gene coordinates

gencode <- read_tsv(
  file.path(
    "/Users/da1078co/Documents/Data/GENCODE/",
    "gencode.v47lift37.basic.annotation.gtf.gz"
  ),
  comment = "#",
  col_types = "cccnn-c-c",
  col_names = c(
    "CHROM", "source", "gene_type",
    "start", "end", "strand", "addinfo"
  )
) |>
  filter(gene_type == "gene") |>
  select(-gene_type) |>
  mutate(
    CHROM = gsub("chr", "", CHROM),
    gene_id = gsub(
      "^gene_id \"([^\"]+)\";.*",
      "\\1",
      addinfo
    ),
    gene_type = gsub(
      ".*gene_type \"([^\"]+)\";.*",
      "\\1",
      addinfo
    ),
    gene_name = gsub(
      ".*gene_name \"([^\"]+)\";.*",
      "\\1",
      addinfo
    )
  ) |>
  select(-addinfo)
head(gencode)

### Adding cis regions

trait_tbg <- trait_tb0 |>
  filter(trait %in% gene_traits) |>
  inner_join(gencode, by = c("trait" = "gene_name")) |>
  select(trait, CHROM, start, end)

trait_tbg

## Other traits

trait_tbng <- trait_tb0 |>
  filter(!trait %in% gene_traits)
trait_tbng$trait

### GWAS sources

| Acronym      | Trait                            | GWAS source                                               |
|--------------|----------------------------------|-----------------------------------------------------------|
| BasalISR     | Basal insulin secretion rate     | IVGTT GWAS from Wood et al 2017                           |
| Clins        | Insulin clearance                | Not found                                                 |
| Clinsb       | Basal insulin clearance          | Not found                                                 |
| Glucagonmin0 | Fasting glucagon                 | Not found                                                 |
| Glucose      | Fasting glucose                  | MAGIC consortium - Chen et al 2021                        |
| GlucoseSens  | Glucose sensitivity              | xinsdG30 from Madsen et al 2024                           |
| HDL          | High-density lipoprotein         | HDL from GLGC - Graham et al 2021                         |
| HOMA_IR      | Insulin resistance index         | MAGIC consortium - Dupuis et al 2011                      |
| HbA1c        | Glycated hemoglobin              | MAGIC consortium - Chen et al 2021                        |
| Insulin      | Fasting insulin                  | MAGIC consortium - Chen et al 2021                        |
| LiverFat     | Liver fat content                | Liu et al 2021                                            |
| OGIS         | Oral glucose insulin sensitivity | Not found                                                 |
| PancFat      | Pancreatic fat content           | Liu et al 2021                                            |
| SAT          | Subcutaneous adipose tissue      | Liu et al 2021                                            |
| TG           | Triglycerides                    | GLGC consortium - Graham et al 2021                       |
| TotGLP1min0  | Fasting GLP-1 levels             | Not found                                                 |
| TwoGlucose   | OGTT two-hour glucose            | MAGIC consortium - Chen et al 2021                        |
| TwoInsulin   | Insulin fold-change post-challenge            | MAGIC consortium - Williamson et al 2023                        |
| VAT          | Visceral adipose tissue          | Liu et al 2021                                            |

Based on this table, only including traits for which we found GWAS summary statistics:

tormv <- c(
  "Clins", "Clinsb", "Glucagonmin0", "OGIS", "TotGLP1min0"
)

trait_tbngf <- trait_tbng |>
  filter(!trait %in% tormv)

URLs used and download process are described in `03_Download_GWAS.sh`

## New table with edges to test

edge_tbf <- bind_rows(
  # Testing links in both directions
  tibble(
    node1 = edge_tb0[, 1, drop = TRUE],
    node2 = edge_tb0[, 2, drop = TRUE]
  ),
  tibble(
    node1 = edge_tb0[, 2, drop = TRUE],
    node2 = edge_tb0[, 1, drop = TRUE]
  )
) |>
  # Only nodes for which we have GWAS summary statistics
  filter(
    !node1 %in% tormv,
    !node2 %in% tormv
  ) |>
  # Specify type of analysis that should be carried out
  mutate(
    atype = ifelse(node1 %in% gene_traits, "SMRHEIDI", "IVWMR")
  ) |>
  # Adding chromosome and position information for gene traits
  left_join(trait_tbg, by = c("node1" = "trait"))

edge_tbf

## Saving

write_tsv(
  edge_tbf,
  file.path(projfold, "data", "edge_table_totest.tsv")
)
