==> scripts/Daniel/scripts/03b_Process_pQTLPanel.ipynb <==
# pQTL panel

library(readr)
library(dplyr)
library(tidyr)
library(purrr)

projfold <- "/Users/da1078co/Documents/Lund/PhD/Projects/BN"

## Olink protein map

Note: gene start and end coordinates are in hg38.

olinkmap <- read_tsv(
  file.path(projfold, "data", "olink_protein_map_3k_v1.tsv"),
  show_col_types = FALSE
)
head(olinkmap)

## UKB-PPP file map

olinkfilemap <- read_tsv(
  file.path(projfold, "data", "olinkfilemapUKB.tsv"),
  show_col_types = FALSE
)
head(olinkfilemap)

Extractin Olink ID

olinkfilemap <- olinkfilemap %>%
  mutate(OlinkID = gsub(".*_(OID[^_]+)_.*", "\\1", name))
head(olinkfilemap)

## Edge table

edge_tbf <- read_tsv(
  file.path(projfold, "data", "edge_table_totest.tsv"),
  show_col_types = FALSE
)
head(edge_tbf)

## Query table

protquery_ukbppp <- edge_tbf |>
  filter(atype == "SMRHEIDI") |>
  select(
    HGNC.symbol = node1,
    start_hg19 = start, end_hg19 = end
  ) |>
  unique() |>
  inner_join(olinkmap, by = "HGNC.symbol") |>
  inner_join(olinkfilemap, by = c("OlinkID"))
head(protquery_ukbppp)

unique(protquery_ukbppp$HGNC.symbol)

## Saving query table

write_tsv(
  protquery_ukbppp,
  file.path(projfold, "data", "protquery_ukbppp.tsv")
)
