==> scripts/Daniel/scripts/07_EdgeResults.ipynb <==
# Results from edge testing

## Libraries

library(readr)
library(dplyr, warn.conflicts = FALSE)
library(tidyr)
library(purrr)
library(ggplot2)

## Plot resolution

options(repr.plot.res = 300)

## Working directory

projfold <- "/Users/da1078co/Documents/Lund/PhD/Projects/BN"

## Table of edges tested

edge_table_totest <- read_tsv(
  file.path(projfold, "data", "edge_table_totest.tsv"),
  show_col_types = FALSE
)
head(edge_table_totest)

## Importing results

edge_table_results <- edge_table_totest |>
  mutate(
    RESOBJ = map2(
      node1, node2,
      ~read_rds(
        file.path(
          projfold, "data", "EdgeTests",
          paste0(.x, "_", .y, ".rds")
        )
      )
    )
  )
print(edge_table_results)

## IVW-MR results

ivwmr_res <- edge_table_results |>
  filter(atype == "IVWMR") |>
  transmute(
    node1, node2,
    resdf = map(RESOBJ, "resdf"),
    failure = map_chr(RESOBJ, "failure")
  ) |>
  unnest(resdf, keep_empty = TRUE)
head(ivwmr_res)

### Top hits

ivwmr_res |>
  filter(PVAL_IVW < 0.01, PVAL_B0Egger > 0.01)

### Saving

write_tsv(
  ivwmr_res,
  file.path(projfold, "data", "ivwmr_res.tsv")
)

## SMR-HEIDI results

smrheidi_res <- edge_table_results |>
  filter(atype == "SMRHEIDI") |>
  transmute(
    node1, node2,
    resdf = map(RESOBJ, "resdf"),
    failure = map_chr(RESOBJ, "failure")
  ) |>
  unnest(resdf, keep_empty = TRUE)
head(smrheidi_res)

### Top hits

smrheidi_res |>
  filter(pval_smr < 0.01, pval_heidi > 0.01)

### Saving

write_tsv(
  smrheidi_res,
  file.path(projfold, "data", "smrheidi_res.tsv")
)
